requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
arsenic==21.8
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
free-proxy==1.0.6
frozenlist==1.3.1
idna==3.3
lxml==4.9.1
multidict==6.0.2
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
structlog==20.2.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
webdriver-manager==3.8.3
yarl==1.8.1
